Rewrite after feedback
Change title and name of the feature to emphasize the main
properties of the proposal.

Write less about interpolation and other features, focus on
indentation and the absence of escape sequences.

Clarify syntax details, why the suggestion is about
a new string format and how it must be designed
to work as a documentation attribute.

Remove the suggestion about using trailing \ for line
concatenation.  The string being verbatim is more important.
For example original Markdown has the opposite definition
of a trailing \.